+++ src/assets/ba_data/python/babase/__init__.py
@@ -238,7 +238,11 @@ from babase._env import DEFAULT_REQUEST_TIMEOUT_SECONDS
 from babase._net import get_ip_address_type, NetworkSubsystem
 from babase._plugin import PluginSpec, Plugin, PluginSubsystem
 from babase._simpledialog import SimpleDialog
-from babase._stringedit import StringEditAdapter, StringEditSubsystem
+from babase._stringedit import (
+    StringEditAdapter,
+    StringEditKind,
+    StringEditSubsystem,
+)
 from babase._text import timestring
 from babase._workspace import WorkspaceSubsystem
 
@@ -439,6 +443,7 @@ __all__ = [
     'SpecialChar',
     'storagename',
     'StringEditAdapter',
+    'StringEditKind',
     'StringEditSubsystem',
     'supports_max_fps',
     'supports_vsync',